[JIMT][Dance AI Modal refactoring] (#55089)
* jimt/dance-ai-modal-refactoring - broke up utils.ts file; move types to ai/types

* jimt/dance-ai-modal-refactoring - pulled all functions out of DanceAiModal

* jimt/dance-ai-modal-refactoring - changed types import

* jimt/dance-ai-modal-refactoring - changed types import

* jimt/dance-ai-modal-refactoring - renaming of methods

* jimt/dance-ai-modal-refactoring - validateAndSetFieldValue now accepts a logger

* jimt/dance-ai-modal-refactoring renamed logger -> invalidValueLogger

* jimt/dance-ai-modal-refactoring - renamed ai/types -> ai/ai-types

* jimt/dance-ai-modal-refactoring - retooled to remove field param in favor of logValues

* jimt/dance-ai-modal-refactoring - restores field key

* jimt/dance-ai-modal-refactoring - moved ai-types back to types

* jimt/dance-ai-modal-refactoring - message/value non-optional and exported
