+++ apps/src/dance/ai/utils/getEmojiImageUrl.ts
@@ -1,0 +1,3 @@
+export const getEmojiImageUrl = (id: string) => {
+  return `/blockly/media/dance/ai/emoji/${id}.svg`;
+};
